+++ packages/moonstone/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## [1.0.0-alpha.5] - 2016-12-16
+
+No changes.
+
 ## [1.0.0-alpha.4] - 2016-12-2
 
 ### Added